projects/cymbal-preproc/src/llm_cr_preproc_requirements.txt
fastapi
hypercorn
langchain
unstructured
pydantic
shapely==1.8.5
faiss-cpu
google-cloud-aiplatform
google-auth
google-cloud-logging
tabulate
pdf2image